+++ pywwt/jupyter_server.py
@@ -1,4 +1,7 @@
 import os
+import json
+from hashlib import md5
+from tornado import web
 from notebook.utils import url_path_join
 from notebook.base.handlers import IPythonHandler
 
@@ -6,20 +9,54 @@ __all__ = ['load_jupyter_server_extension']
 
 
 STATIC_DIR = os.path.join(os.path.dirname(__file__), 'nbextension', 'static')
+CONFIG = os.path.expanduser('~/.pywwt')
 
 
-class WWTHTMLHandler(IPythonHandler):
-    def get(self):
-        with open(os.path.join(STATIC_DIR, 'wwt.html')) as f:
+class WWTFileHandler(IPythonHandler):
+
+    def get(self, filename):
+
+        filename = os.path.basename(filename)
+
+        # First we check if this is a standard file in the static directory
+        if os.path.exists(os.path.join(STATIC_DIR, filename)):
+            path = os.path.join(STATIC_DIR, filename)
+        else:
+            # If not, we open the config file which should contain a JSON
+            # dictionary with filenames and paths.
+            if not os.path.exists(CONFIG):
+                raise web.HTTPError(404)
+            with open(CONFIG) as f:
+                config = json.load(f)
+            if filename in config['paths']:
+                path = config['paths'][filename]
+            else:
+                raise web.HTTPError(404)
+
+        with open(path, 'rb') as f:
             content = f.read()
+
         self.finish(content)
 
 
-class WWTJSHandler(IPythonHandler):
-    def get(self):
-        with open(os.path.join(STATIC_DIR, 'wwt_json_api.js')) as f:
-            content = f.read()
-        self.finish(content)
+def serve_file(path, extension=''):
+
+    if not os.path.exists(path):
+        raise ValueError("Path {0} does not exist".format(path))
+
+    hash = md5(path.encode('utf-8')).hexdigest() + extension
+
+    with open(CONFIG) as f:
+        config = json.load(f)
+
+    if hash not in config['paths']:
+
+        config['paths'][hash] = os.path.abspath(path)
+
+        with open(CONFIG, 'w') as f:
+            json.dump(config, f)
+
+    return '/wwt/' + hash
 
 
 def load_jupyter_server_extension(nb_server_app):
@@ -27,8 +64,10 @@ def load_jupyter_server_extension(nb_server_app):
     web_app = nb_server_app.web_app
     host_pattern = '.*$'
 
-    route_pattern = url_path_join(web_app.settings['base_url'], '/wwt.html')
-    web_app.add_handlers(host_pattern, [(route_pattern, WWTHTMLHandler)])
+    if not os.path.exists(CONFIG):
+        config = {'paths': {}}
+        with open(CONFIG, 'w') as f:
+            json.dump(config, f)
 
-    route_pattern = url_path_join(web_app.settings['base_url'], '/wwt_json_api.js')
-    web_app.add_handlers(host_pattern, [(route_pattern, WWTJSHandler)])
+    route_pattern = url_path_join(web_app.settings['base_url'], '/wwt/(.*)')
+    web_app.add_handlers(host_pattern, [(route_pattern, WWTFileHandler)])