+++ CHANGES.rst
@@ -1,6 +1,8 @@
 0.4.0 (unreleased)
 ------------------
 
+- Added ``load_fits_data`` method for Qt and Jupyter clients. [#78]
+
 - Fix compatibility with Jupyter Lab. [#63, #65]
 
 - Added GUI controls for imagery layers. [#64]